Add basic GOVUK page layout
Here we begin to use the GOVUK design system.

We add partials for the header and footer, update the applicaiton layout
to include the required elements and add the assets to the asset
pipeline that we need to use in the view (rather than in the sass
files).

It is really important to include the `js-enabled` class to the body
element if we want the GOVUK JS to work later.