Add GOV.UK Frontend Javascript
Here we take the necessary steps to load the GOV.UK JS into the assets
pipeline and initalise any components on our pages.

Right now we are not worried about transpilling the JS, we think the JS
is compatible with our target browsers [1] but we will need to test
that.

We are also not worrying about minification right now, this is all
mainly due to using Rails 7 with it's new asset pipeline.