Ensure that the email addresses are lowercase
We match authenticated users to our registered users by email address,
so we don't want issues of case sensitivity. Here we downcase when
creating registered users in the applicaiton.

We know that the Microsoft sign in returns emails with case issues, mine
comes back as `meyric.RAWLINGS@` - being consistent across the app helps
to mitigate this.