Do not compress sass in test environment
Without this option the asset pipeline appears to try and compress the
sass files and will fail. You can see this happen by removing this line,
ensuring you have no precompiled assets with `bin/rails assets:clobber`
and running the specs `bin/rails spec`.

In development, the assets are compiled on the fly and this does not
seem to be an issue.

In production, we precompile the assets and this does not appear to be
an issue.

I'm afraid I can shed no further light on this other than saying I hate
the asset pipeline.