Add Projects model
This is (at least for now) a very simple model which only stores our internal ID (as a UUID) and the URN (Unique Reference Number) of the school which is converting.

As a note, the URN is _actually_ an integer (and is represented as one in the Academies API), not just a string of numbers.