Switch Omniauth strategy
We realised that ideally we want to work in Azure Active Directory
single tenant mode i.e. only users in the DfE directory can auth.

The Omniauth strategy we were using did not support this mode so here we
swicth to one that does.

The gem we chose is a fork of the original Microsoft one that has been
updated so seems like a good choice.

This work also adds an additional environment variable:

AZURE_TENANT_ID

which we will document next.

We are also checking fot the required environment variables in the
Dotenv initializer. This is a quality of life thing for developers,
booting the application with one of these missing will raise an clear
error. This is less useful in the test environment, but is an
acceptatble side effect of the setup.