Extend the Academies API wrapper
We have our base connection class, here we:

Create an Establishment model, this will wrap a single establishment
(think school) that is returned from the API.

We extend the Client to allow it to fetch an establishment.

Add basic error handling by returning a simple Result object that
contains either an error or the Establishment object for use, this
should allow us to work with the result and help the user understand if
something goes wrong by surfacing errors.

When we cannot connect to the API we raise an error, I am not sure what
we want to do if this happens, but it feels significant enough to raise
an error.