Refactor authenticaitons helpers
In our request specs We discovered our authentication mocks were not
working as we expected, the user was not technically signed in.

Here we create new sign in helpers that are focused on mocking a
successful authentication and an unsuccessful one. We update the specs
to use the new helpers and crucially, stub out the `user_id` method on
any contoller we use as this is what constitutes a sign in.

Right now I cannot find a good way to stub user_id on any controller
and I am a bit iffy on using `allow_any_instance`, but fixing this here
unblocks other work.

This is only an issue in request specs as features excercise the whole
system and create a session.