Add YAML conversion workflow
Co-authored-by: Nick Jackson <nick@dxw.com>

We want to generate task list sections and tasks based on a YAML
definition. This will mean that we can easily change and update the task
list, as well as introduce new definitions for other workflows.

This also adds the `ajv` package so that we can validate that our
workflows fit the schema defined in `./workflows/schemas/task-list.json`