Add pundit policy for creating new projects
This policy restricts creation of new projects to team leaders, by
checking that permissions exist for users trying to view or submit the
new project form. It also removes the "new project" button from the
index view for users who are not team leaders.

A new `team_leader` trait is made available on the `User` factory to
simplify tests of this condition.