Scope projects index to only user's own projects
Use the pundit policy to scope the projects displayed on the home screen
to only the ones where the user is assigned as the delivery officer,
unless the user is a team leader (in which case they can view all
projects).

To keep existing tests passing, tests for being able to access a project
from the home page are modified to use a team leader. More comprehensive
testing of this behaviour will be added once authorisation is
implemented for the Project `show` view.