Update Workflow schema with Action
Here we add the actions as an array with the `title` property.

Actions is required and must contain at least one element.

We remove some unused attributes from the schema.

We also added the workflow validation as a run script so it lives with
other that are in a similar space.