Update TaskListCreator service
This now adds the actions for a task, we do so using `insert_all` for a
tiny performance gain as we are at the bottom of the hiearchy and have
the parent task available.

We also refactor to use `fetch` which is slightly more defensive as we
can provide a value should there be no Tasks or Actions, even though
this is not something we expect.