Update Conversion workflow with latest content
We have worked with design and user research to come to our first
version of the content for the Conversion task list.

We expect this to change, this is our first opportunity to add value and
engage users by helping them manage their projects with the existing
processes.

This content is copy and pasted for the Lucid board[1], we've only
corrected typos when we saw them, further refinements can be made to
this file without impact any projects that are already in the
application.

[1] https://lucid.app/lucidspark/76803306-efd1-4188-a83b-7afa139655b4/edit?view_items=j9JwWBjAVpOU&invitationId=inv_02b61ee8-9a4e-46c8-b1fe-6129fe53530b#