Update how we define API models
Currently, we pass the JSON response from the API into the constructor
of our API models. However, this doesn't work nicely with FactoryBot,
where we have to use the `initialize_with` method in order to create
mock objects.

Use the `ActiveModel::Serializers::JSON` along with an `attribute_map`
in order to map API response property names to the desired attribute
names. Abstract it to a base class so we can keep the API models DRY.