Validate a single matching conversion project exists on project creation
When creating a project, catch any exceptions around either missing or
multiple conversion projects from the API and present an in-form error
message to the user.

Simplifies places in the feature specs where the
`mock_successful_api_establishment_response` helper is called with the
new `mock_successful_api_responses`, which mocks both establishment and
conversion project calls in one go.