Update migrations for MS SQL Server
The `enable_extension` method is part of the Postgres adaptor that we
are moving away form and would cause an error on the MS SQL Server.

We can remove it now as from this point forward, migrations will run
against an MS SQL DB, any developer with a Postgres database will have
to setup the new one and we have no production or staging environments
to worry about.