Update Database environment
We have to manage the MS SQL Server in a different way.

Here we are moving as much configuration to `database.yml` as we can and
overriding the values from environment variables.

With the connection config in place, we can run `bin/rails db:prepare`
to create the databases and run the migration, note this will do both
development and test environments.

With the database prepared, the schema is updated for the new adaptor.

We also tied the Dotenv files to try and make things clearer.