Add SQL Server environment
As we have to use MS SQL Server on our development machines, which are
mostly Mac OS, we cannot run the database natively.

We could run everything inside containers, but we prefer to have that as an
option rather than the default.

Running the database in Docker and the application natively offers some
advantages many developers prefer.

Here we add the compose file and environment to launch SQL Server and
have the application connect to it locally.

We hope to automate much of this later, but here we are making sure the
basic setup works and is documented.