Update no-docker scripts
These scripts let a developer start a local development environment
quickly and simply.

Whilst this is the 'no-docker' path, we now have to use Docker for the
SQL Server as there is no native Mac OS version - the naming is
unfortunate, but not worth updating at this point - we only run SQL
server in a container, the applicaiton is running natively.