Move authentication to be included in ApplicationController
This change means that all controller actions are now automatically
authenticated unless explicitly instructed otherwise using
`skip_before_action` (as in `SessionsController`).

In the case of `SessionsController` we use `only` to explicitly list
actions which should be excluded from authentication. Although in this
case we list every action, it means that if we add a new action to the
controller it will be protected unless we make the conscious decision to
also exclude it.