Refactor tests now that team leader ID is a non-nullable field
We have added a non-nullable association to `team_leader` on the
`project` record.

Update the project factory with a team_leader association. This uses a
random email address so that when we create multiple `project`s in a
spec, we don't see a record not unique error (due to the unique
constraint on the email column in `user`).

Some specs were written before we started to use FactoryBot. Update
these to use the `project` factory for the above reason.