Handle nested responses in the BaseApiModel
Some of the Academies API responses contain nested data. We have decided
to surface these on our models, for example, mapping
`establishmentType.name` to a `type` attribute on
`AcademiesApi::Establishment`.

Update the BaseApiModel to handle this. We can provide the full access
path in dot notation as the value in the `attribute_map`.