Swap out the conversion project endpoint for the trust endpoint
We want to decouple the application from prepare for conversions. Currently,
if there is no conversion project available, our app won't function - users won't
be able to create a project. 

We want to use the Academies API trust endpoint to collect trust information -
which we currently get from the conversion project.

This removes the ConversionProject model, and adds a Trust model.  In order to
get a trust we need to introduce a new field of trust UKPRN to the new project
form. We will no longer be able to get proposed opening date, so we will also need
to introduce a field for that - which will come in the next commit.

This includes a small refactor of our feature specs, removing project
validation failed tests - we have agreed that we can maintain a more balanced test suite
by relying on our model specs for this. The `rails-controller-testing` Gem is
needed in order to use the `render_template` method.