Add target completion date to Project
Previously, we pulled this in from the Conversion Project endpoint. Now that
we've removed this, we need to ask the user for the target completion date when they
create the Project. 

Because the target completion date is always on the first of the month, remove
the day field from the date input, and add some validation around this.