Allow users to view previous project notes
Users must be able to see the previous notes for a project, including the name
of the author (email for now), the date the note was created, and the body of the
note. 

Order notes in descending order by `created_at`, so that the most recent notes
are first. Add the `ActiveSupport::Testing::TimeHelpers` module to make testing
this easy.