Tidy up partials used by a single view file
It seems typical that when a partial is only used by a single view, that it is
nested in a folder with the same name. E.g., the putting the project
information side navigation used on the `show` view in
`project_information/show/side_navigation`. 

Tidy up any partials used by a single view file in this manner.