Show additional information on the project index page
We want to show the school type, target completion date, trust name, and local
authority name for each project on the index page as in the prototype. This
also removes the assigned delivery officer.

There is no component which matches how we want this to be displayed, so I have
created it myself. Attempting to DRY it up any further makes things more
difficult to follow because of the layout classes.

This also renders the projects index list as a `ul` for easier testing. This
seems to be more semantically correct anyway.