Move the new note form onto its own page
Currently, the new note form is underneath the project summary and
subnavigation. This means that the error summary sits halfway down the page,
whereas the GOV.UK guidance says it should sit at the top.   

After a chat with our interaction designer, we have decided to move the new
note form onto its own page. This solves the error message issue, and has an
additional benefit that editing messages will be easier if we choose to
implement that in the future.

This includes a few style updates:
- Add a cancel link to the new note form
- Rename the textarea label from "Notes" to "Enter note"
- Change the notes form submission button from "Add note" to "Save note"
- Remove the top separator of the list of notes