Update Clear legal documents section
Through user research we came to what we think is a fair order and
naming for this section.

This work simply updates the conversion workflow file to reflect these
changes.

As we have no production data, there is no concerns for existing
Projects not reflecting this change.

As we also decided that all tasks in this section will have the same set
of Actions for our MVP, we updated any that were not already the agreed
shape of:

- Received
- Cleared
- Saved in SharePoint