Render hints and guidance
For now, render the hint as part of the checkbox, and conditionally render the
guidance.

The guidance (rendered as a details component) currently sits within the
checkbox fieldset (because we have modelled actions checkboxes within a
fieldset). I can imagine this will change as move move toward each action being
a distinct thing.