Update the workflow file and schema
Now that the Task model can be optional, we want to be able to build
optional Tasks from the workflow file.

We have false as a default at the database level and here by using
`fetch` with a default of false - this means if the `optional` attribute
is not included in the yaml optional will be set to false.