Users can mark optional tasks as not applicable
With all the mechanics in-place, we can add the UI to let users mark
optional tasks as 'not applicable'.

The view is a little bit odd here because we have multiple objects and
type of objects acting as one set of GOV.UK checkboxes. This is
inevitable as we are working at the edge of that the component is
designed to do, we might want to come back and create a new component
with a cleaner interface later. For now, this works.

A set of GOV.UK checkboxes can normally be 'exclusive'[1] in that when
the exclusive option is set, the other options are unset - we replicate
this behaviour without javascript by setting all the Tasks Actions
completed state to false when submitted as not applicable.

We renamed the strong params method to `task_params` as it felt more
descriptive.

[1] https://design-system.service.gov.uk/components/checkboxes/