Update the workflow and TaskListCreator to handle hints and guidance
The TaskListCreator and workflow need updating to handle task-level hints and
guidance. Update them, following the same structure as used for the
action-level hints and guidance.

Because some of the keys of the workflow task are now optional, we must provide
a default value to `fetch` for when the key cannot be found. Add tests to check
explicitly that workflows with empty optional values work as expected.