Remove User null constraints
We are about to change which users can create new projects, going from a
team lead to a regional delivery officer.

This has the knock on effect of meaning that projects will exist with
these two ids not present in the database until they are assigned by the
appropriate user or automatically.

We'll remove the null constraints on the database and validations to
make the changes and look to re-introduce the appropriate one later if
required.