Change project create policy
Here we are switching the type of user who can create new projects from
a team lead to a regional delivery officer.

As only regional delivery officers can create projects, we know that the
current user will be a regional delivery officer and can assign that
user into the `project.regional_delivery_officer` association, like we
used to for team leads.

Now that the regional delivery officer is assigned automatically there
is no need to show the select box in the interface.