Team lead is assigned on case worker assignment
Previously the team lead was assigned when a project was created as we
expected team leads to create projects.

We then decided that regional delivery officers should create (or
'handover') a project so can no longer assign the team lead that way.

Team leads will always assign the case worker to a project, so we can
use that to assign the team lead, when the casw worker assignment
happens we can use the current use (who must be a team lead) to the team
leader association.

Tidied up the flash messages in the ProjectController as well.