Move create_unassigned_project to a helper
Creating a project requires mocking the Academies API endpoint to
validate the establishment and trust exist. To make this easier we have
this `create_unassigned_project` spec helper that takes care of things
for us.

We would like to use this helper in another spec so we will move it into
a helper.