Change the test coverage around assignment
Users assign other users to roles on projects. We have changed how this
works recently.

Here we make this assignment jounrnet more explicit by changing the
feature spec that covers it to exercise the user jounrney.

We have already introduced request specs to cover off the behaviour in
the deleted spec, checking the visablity of the edit button and the
permissions.