Add a scope for caseworkers
Users who are neither a team lead or a regional delivery officer are
considered caseworkers.

We want to show a list of only caseworkers when assigning them to
projects. Here we add a scope to allow that.

It might make sense to make the caseworker role explicit at some point,
but that can come later.