Add first name and last name to the user model
Currently, we refer to users throughout the app by email. In order to be more
user friendly, we want to use the user's names instead. 

This adds first name and last name to the user model. The migration doesn't
specify that these field are required, so no existing user record will be
impacted.

This also adds "column" specs to the User model specs in order to match the
other models.