Fix: test environment has SENTRY_ENV
The SENTRY_ENV environment variables is required in all instances, this
caught me out when I came to do some work.

We can set the SENTRY_ENV in .env.test as it is not a secret and we can
make the error message should the variables somehow not be set with the
dotenv initalizer.