Show user's full name on the project information tab
Now that the User model has first and last name, we want to display this
alongside email on the project information tab.

This adds email as a `mailto:` link after the full name.

It also adds a helper to DRY up the logic around displaying a name. So that this
is backwards compatible for pre-existing users, we display the email address for
users who don't have a name.