Add diocese details to project information
A diocese is available only on certain types of faith schools.

Here we add the diocese name to the attributes we get on an
establishment from the Academies API and show that on the project
information view.

The default behaviour of showing 'not applicable' may not be the best
for this attribute, but it is good enough for this version.