Update project information layout
We want a little more control on the specs for this view, as each
'section' is hardwired to the view, we add ids which let us target those
details specifically.

As well as a little more control, the big spec is broken up a little
making it slightly more approachable.