Refactor feature spec for redability
I found this feature really hard to parse, the use of the helper felt
like it was trying to do too much.

Generally, I don't feel the need to test the Rails locale system is
working for the labels on this view, the main priority for the user is
to view the values.

Capybara gives us enough to describe the behaviour without haveing to
understand how the helper works on top.