Add advisory board data to Project
The advisory board ultimately approves a conversion with or without
conditions.

Right now it is not clear how this informaiton is used by our users, but we
do have strong evidence to suggest that it is needed.

In a perfect world, we would read this data out of the prepare project, but
our focus is not on automating the handover, so we we collect the
information in the handover from a regional delivery officer.

Here we add the columns to the project table - we know from the handover
proforma document what regardless of the type of project (conversion,
transfer or significant change) this data is always collected.