Rename `trust_ukprn` to `incoming_trust_ukprn`
When we modelled `trust` we kept the naming simple, we've since learnt
that `trust` is a keyword in Ruby in certain conditions[1].

A conversion represents a school becoming an academy by joining a Trust,
we use the term 'incoming trust' to describe to which trust the school is
joining.

Later when adding transfers we will be storing both the
incomming and outgoing trust i.e the new trust and the trust the school
is leaving - changing this column name now solves the Ruby name clash and
sets us up nicely for a time when we introduce the concept of
`outgoing_trust_ukprn` as well.

[1] https://ruby-doc.org/core-3.1.2/Object.html#method-i-trust