Introduce a UKPRN validator
A United Kingdom Provider Reference Number is a unique number allocated
to providers after successful UK Register of Learning Providers [1]
registration. This is an 8-digit number that always start with a 1.

Both establishments and trusts potentially have a UKPRN so it is a good
candidate for use as an identifier in our service.

Here we add a custom validator.

[1] https://www.ukrlp.co.uk/