Show the advisory board data
Initially this is displayed on the project information tab.

We introduce a presenter[1] here to make dealing with how to render the
values simpler, right now we will have Projects without a advisory board
date, so we want to handle that case (as well as being defensive and
guarding against nil values).

[1] https://www.rubyguides.com/2019/09/rails-patterns-presenter-service/