Refactor TaskListCreator to accept workflow root as parameter
This change makes it easier to separate workflows into multiple files by
clearing up some test behaviour, and allowing us to reduce the amount of
mocking needed.

This also includes some alterations to the exact behaviour which is
being tested in the create projects feature spec; previously it mocked
the YAML which was being loaded at project creation time, it now uses
the real definition. This does mean that if the definition changes we
may have to modify the test, but it also means we're doing less
modification of the system whilst under test.