Split workflow definition into one file per section
The current workflow file is getting large and unwieldy, making it
harder to safely edit and introducing more chance of things like merge
conflicts. To resolve this, split it down into one file per section,
with a new master definition file which specifies the sections in order.

There are no functional changes as part of this.