Add environments to GitHub workflows
We are trying to be as public as possible with out progress, by adding
this `environment` attributes, these workflows should start to show
what code is running in which environment and build a history of
deployments in GitHub.

Right now these workflows do not take the last step to actually deploy
the code, but that work will come and this is good enough for the
moument.