Fix: Authorization and scopes
Passing the Note class to `authorize` will silence the
`verify_authorized` warning, but it is not a good idea, always pass a
instance to authorize, if we actually did anything in the policy this
would blow up.

As all users can view all notes, we can add a Policy scope, this keep
things in the right concerns, the scope returns all notes regardless of
the user and then we filter that set down to either a projects notes or
tasks notes.

Now we have a policy scope, we turn off `verify_authorized` on the index
action and explicitally cover that with `verify_policy_scoped`.

We can also remove the `index` policy action as this is handled by the
new scope.

Finally we can tidy up the `project_level_notes` scope on Note by having
it take a argument that is the project we want to see notes for.

All this behaviour is actually covered in our specs, but this refactor
makes things a little more robust.