Compose depends_on includes a condition
We intermittently see our test failing in CI because the MSSQL is not
ready to accept connections when Rails attempts to create the test
database.

According to the Docker Compose documentation [1]
you can supply a condition to each depends_on.

It's not clear what `service_completed_successfully` actually is, but I
feel like it is a quick win if it does make our CI a little less flakey.

The real challenge seems to be

> what constitutes ready to MSSQL server?

And there does not seem to be simple answer we can use in this context.

[1] https://docs.docker.com/compose/compose-file/#depends_on