MSSQL Docker compose healthcheck
We have seen lots of intermittent failures in CI that we believe are due
to the MSSQL server container not starting before the applications
attempts to utilise it.

This results in an error:

> TinyTds::Error: Unable to connect: Adaptive Server is unavailable or
does not exist (test-db)

To avoid this we want the applicaiton to start once the MSSQL server is
ready to accept connections.

We tried a number of ways to do this and settled on what is here.

`depends_on` can use a `healthcheck`[1] to make sure the containers are
considered healthy before starting the dependent container.

We have to use a seperate container to run the `sqlcmd` tool to attach
to the running MSSQL server, if we can attach, we consider ourselves
healthy and so achieve the goal. This is because the `sqlcmd` is not
included in the MSSQL server image.

With a way to confirm the database is ready, we make the applicaiton
container dependent on the health container.

The helathcheck runs 10 times every 2 seconds giving us a 20 second
window for the MSSQL server to become ready.

[1] https://docs.docker.com/compose/compose-file/#healthcheck