Paginate the projects index page
For now, we send a request to the API for each project on the project index
page. These requests are very slow, resulting in a very, very slow page load time
when there are multiple projects. We want to paginate the projects in order to
reduce the load.

This paginates the projects index page. It also moves the `Add a new project`
button to the top of the page, because it was looking a bit odd next to the
pagination component.