Switch database container image for ARM macs
The Microsoft container image for MSSQL Server[1] does not support ARM
Macs. As some of the team use ARM Macs, we want to be able to support
them.

Whilst there are multiple issues open[2] to resolve this, there is not
idication that it is happening any time soon.

Our only option it to use the 'Azure SQL Edge' image. Whilst this is not
strictly a MSSQL image, it is a product aimed at the IOT market[3], it
contains MSSQL and appears to have all the feaures we will ever use[4].

[1] https://hub.docker.com/_/microsoft-mssql-server?tab=description
[2] https://github.com/microsoft/mssql-docker/issues/734
[3] https://learn.microsoft.com/en-us/azure/azure-sql-edge/overview
[4] https://learn.microsoft.com/en-us/azure/azure-sql-edge/features?source=recommendations