Remove scripts for running in Docker locally
The development team does not use this functionality and the scripts are
out-of-date.

Here we remove the scripts and related files for running the application
locally with Docker to prevent any future confusion.

The Application database, MSSQL must still be run with Docker,
everything else can be run natively via the 'scripts to rule them all'
pattern documented.

The Application can also be run 'hybrid' whith the backing services
(MSSQL and Redis) running in Docker and the applicaiton natively.