Split "action" into its own schema definition file
Since the "action" object is starting to become more complex and carry
additional options, move it into its own definition file so that our
schemas are smaller, more focussed and more easily understood.

This does not change the structure or validity of any of the workflow
files; instead it changes how we organise the definition of what a
workflow file should look like.