Make projects index summary responsive
Currently, the projects index summary wraps onto two lines when overlaps occur
on small viewports and doesn't look great. We want to ensure the summary looks
good on any screen size.

This applies responsive styling to the summary when the viewport is smaller than
desktop size. Unfortunately, this was difficult to achieve with the GOV.UK grid
system, so instead uses flexbox.

This includes a general tidy up of the component, which defines a
`govuk-section-break--s` class that is absent from the GOV.UK Frontend styles.