Add DateTime formats
We are storing significant events such as a project being assigned and
closing as DateTimes, we don't want to show them to users with the time.

Here we add date formats for showing a DateTime as just a date and also
as a date with time for completeness.

We'll use these when we want to show the users any of the significant
dates.