Actions now submit state individually
Previously, the state of all actions was submitted as a single fieldset
with an array of checked values. This commit makes a subtle change to
the behaviour so that each action is submitted as an individual field
(assuming it is checked).

This means that we can better isolate the behaviour of each action in
the form. Initially this is to support improvements to form layout, but
in future this will also make it easier for some actions to have
non-single-checkbox behaviours, such as text inputs, radios or multiple
choice.

The `task_params` method is removed to simplify the flow of the form
logic, since we are no longer attempting to mass assign values.