Add errors when project dates are invalid
Currently, when an invalid date such as 42,01,2022 or -1,01,2022 is set, we see
a MultiParameter error.

Unfortunately, there doesn't seem to be any particularly nice way to fix this.

This overrides the setter for the Project date attributes. We rescue from an
invalid dates, and add a custom error. Simply setting the error in the `rescue`
won't work as the date is nillfied and we will see the error associated with
the `presence` validation. 

This solution does not show the user their previous incorrect input. However,
to enable that does not seem worth the effort at this point.