Remove `Signed by secretary of state` from `Direction to transfer `
We have had some feedback that the `Signed by secretary of state` section is not
needed in the `Direction to transfer` task.

The caseworker doesn't need to arrange for these documents to be signed by
secretary of state - academy operations does this for them, after they send
these documents to them with the single worksheet.