Add type column to project
This is a new column to store the type of a project. It's named
`project_type` since `type` is reserved for Rails's single table
inheritance.

This uses an enum to explicitly list the possible project types, since
this has implications for loading the workflows and we want to try keep
it non-brittle.

Co-authored-by: Laura Porter <laura@dxw.com>