Revert "Add errors when project dates are invalid"
The development environment is currently broken, when creating a new
project we see the following errors:

```
2 error(s) on assignment of multiparameter attributes [error on
  assignment [1, 1, 2023] to target_completion_date (super: no
  superclass method `target_completion_date=' for #<Project > Did you
  mean?  target_completion_date_is_in_the_future),error on assignment
  [1, 10, 2022] to advisory_board_date (super: no superclass method
  `advisory_board_date=' for #<Project >)]
```

This behaviour cannot be replicated locally, right now our next best
step is to remove this work and test in development.

Later work moves the multiparameter validation introduced here, so
ruling this out now would be helpful.

This reverts commit f05b64bbc21e649bb67811e5bab5ee816f8c3b86.