Configure simple cov to output json
We would like Sonarcloud to pick up our coverage, according to this:

https://docs.sonarcloud.io/enriching/test-coverage/test-coverage-parameters/

Outputting json is required so this configures simple cov.

At this point, I have not idea how sonarcloud can read the file, but
this might be a 'it just works' things so we are happy to try as this
has no other impact that we can see.