Introduce a create project form object
Form objects is a design pattern. The are objects that sit between a
form and the model(s) that back it.

As with most decisions, there are trade offs; Here using a
form object will give us:

- a context to validate a new project in
- a place to deal witht the multiparameters from dates in the form
- a different option to handle creating multiple objects from one form,
  a Project and Note in this case
- a place to trigger any notifications
- a way to thin out the controller

Which all feel valuable.

We are only introducing the object in this commit.

We clearly define what attributes we expect the form to submit and which
validations apply when creating a new Project.

Handling multiparameter dates is tricky, this approach overrides the
setters for date attributes on the form object and validate each
parameter there; This approach is not ideal, but we do not have the
luxuary of time to commit right now.

The user approach is that when all params are empty we behave as
`blank`, if there is a param missing or an invalid param, we treat the
whole date as invalid - this feels like the a good comprimise for our
users.