Set up Selenium webdriver
We are using selenium to help with accessibilty testing. This is a step
in automating our accessibilty testing, with the aim to run the
accessibilty tests within ci, only running against specific tests
focused on this.
Selenium allows a --headless option, which allows quicker running time
of these tests than without the headless option.