Add an open and completed scope to projects
We are going to want to divide the project list on the project index page into
open and completed projects, i.e. projects with a nil `completed_at` date and
projects with a non-nil `completed_at` date.

Add a couple of simple scopes for these separated lists of projects,
plus their tests.